+++ CHANGELOG.md
@@ -9,6 +9,7 @@ New in 0.9.0
 * *Addition* - `phrase` accepts vector durations, representing repeated notes.
 * *Addition* - `phrase` accepts an optional third argument for velocities.
 * *Addition* - `but` replaces part of a melody.
+* *Addition* - `tempo` applies a transformation to both time and duration.
 * *Breaking change* - `where` ignores missing keys.
 
 New in 0.8.0